Rename all mocks
