+++ src/mocks/stateMock.ts
@@ -1,0 +1,20 @@
+import { AuthStatus } from 'features/auth/authSlice';
+import { userWithoutReviewMock, userWithReviewMock } from './userMock';
+
+const status: AuthStatus = 'IDLE';
+
+export const loggedInUserWithoutReviewsStateMock = {
+  auth: {
+    user: userWithoutReviewMock,
+    status,
+    error: null,
+  },
+};
+
+export const loggedInUserWithReviewsStateMock = {
+  auth: {
+    user: userWithReviewMock,
+    status,
+    error: null,
+  },
+};
