--- src/mocks/user.ts
@@ -1,16 +1,0 @@
-import User from 'types/User';
-
-interface UserWithPassword extends User {
-  password: string; 
-}
-
-const user: UserWithPassword = {
-  email: 'correct@email.com',
-  _id: 'testUserIddrfgdfggdfdgdfgdf',
-  firebaseId: 'testFirebaseId',
-  firstName: 'testFirstName',
-  lastName: 'testLastName',
-  password: 'correctPassword',
-};
-
-export default user;
